Add pre-alpha action
This adds everything required to publish the pre-alpha version of
the action.

It probably doesn't do everything we need, and it might not even work,
but it's a start.

This should accomplish the following:

- Dockerfile with everything required to run the CDK
- Script that can run CDK commands
- Workflow that publishes the Dockerfile to Docker Hub
- Action configuration that can be used to execute the Dockerfile

These should all work together fairly nicely to be able to run CDK commands
from a GitHub workflow elsewhere, that will be able to run CDK commands for
a JVM app, something that for some reason isn't very popular elsewhere.